Proposal for introducing capabilities in kpasswdd
KDC daemon already support dropping capabilities (using capng if available).

This patch only moves function kdc/main/switch_environment() to libheimbase/heim_switch_environment().

This way, KDC and KPASSWDD daemons share the same routine to manage privileges.
It also allows KPASSWDD to be run as a non-root user while using standard port.
